updating enum to single file instead of many files in shared pack
