--- shared/enums/account.js
@@ -1,7 +1,0 @@
-const AccountTypes = Object.freeze({
-  INCOME: "income",
-  EXPENSE: "expense",
-  ASSET: "asset",
-});
-
-module.exports = { AccountTypes };
