--- shared/enums/local-storage.js
@@ -1,7 +1,0 @@
-const LocalStorageKeys = Object.freeze({
-  ACCESS: "sims-access",
-  REFRESH: "sims-refresh",
-  CURRENT_USER: "sims-current-user",
-});
-
-module.exports = { LocalStorageKeys };
