--- shared/enums/user.js
@@ -1,9 +1,0 @@
-const UserType = Object.freeze({
-  ADMIN: "admin",
-  MANAGER: "manager",
-  SALES_MAN: "sales-man",
-  INVENTORY: "inventory",
-  CASHIER: "cashier",
-});
-
-module.exports = { UserType };
